[BACKPORT] Add Debug.getRss()
This change adds the getRss() method, which will return RSS (as
estimated by /proc/<pid>/status) and memtrack graphics, memtrack gl, and
memtrack other. This is intended to be used a drop-in replacement for
getPss() where efficiency is a higher priority than accuracy. Unlike
getPss(), this method does not provide the option for collecting USS or
SwapPss as these are not emitted in status.

There are some formatting changes included here that were required for
'repo upload'.

Test: Added method call in CTS's DebugTest; included unit test for
      backing ProcMemInfo::StatusRss method.
Bug: 296454553
Change-Id: I20e7999dce46f6cdf01cb740ab9576937711dfef