PSS fix permission dialog not re-showing after config change
When the permission dialog is showing, and a configuration change
happens (e.g. display size change), there is an issue that the
dialog is not shown again.

The problem seems to be that the permission activity and dialog are
created, but the callback `onDialogDismissed` is immediately triggered,
in the permission activity, which calls finish() on the activity.

Somehow, removing cancel and dismiss listeners in onDestroy, solves
this problem. I was surprised, because I don't understand how removing
a listener on the previous activity can affect the listeners on the
new activity, which uses a new instance of a dialog.

Change-Id: I3e87933f9439ab046eda92bc5863b29fad57f0db
Test: Manual - Share screen using the zoom app. Then change display size multiple times using adb. Not sure how to write an automated test for this yet.
Fixes: 286563020